+++ custom_components/lock_code_manager/manifest.json
@@ -9,7 +9,10 @@
     "@raman325"
   ],
   "config_flow": true,
-  "dependencies": [],
+  "dependencies": [
+    "frontend",
+    "lovelace"
+  ],
   "documentation": "https://github.com/raman325/lock_code_manager",
   "iot_class": "local_polling",
   "issue_tracker": "https://github.com/raman325/lock_code_manager/issues",
