+++ custom_components/lock_code_manager/strategy/src/view-strategy.ts
@@ -1,58 +1,12 @@
 import { ReactiveElement } from 'lit';
+// import { customElement } from 'lit/decorators';
 
-import { DOMAIN } from './const';
-import { ConfigEntry, EntityRegistryEntry, HomeAssistant } from './ha_type_stubs';
-import { compareAndSortEntities, createLockCodeManagerEntity, getSlotMapping } from './helpers';
-import {
-  ConfigEntryToEntities,
-  LockCodeManagerDashboardStrategyConfig,
-  LockCodeManagerEntityEntry,
-  LockCodeManagerViewStrategyConfig,
-  SlotMapping
-} from './types';
+import { HomeAssistant } from './ha_type_stubs';
+import { compareAndSortEntities, getSlotMapping } from './helpers';
+import { LockCodeManagerViewStrategyConfig, SlotMapping } from './types';
 
-class LockCodeManagerDashboard extends ReactiveElement {
-  static async generate(config: LockCodeManagerDashboardStrategyConfig, hass: HomeAssistant) {
-    const [lockConfigEntries, entityEntries] = await Promise.all([
-      hass.callWS<ConfigEntry[]>({
-        domain: DOMAIN,
-        type: 'config_entries/get'
-      }),
-      hass.callWS<EntityRegistryEntry[]>({
-        type: 'config/entity_registry/list'
-      })
-    ]);
-    const lockConfigEntryIDs = lockConfigEntries.map((configEntry) => configEntry.entry_id);
-    const lockEntityEntries: LockCodeManagerEntityEntry[] = entityEntries
-      .filter((entity) => lockConfigEntryIDs.includes(entity.config_entry_id))
-      .map((entity) => createLockCodeManagerEntity(entity));
-
-    const configEntriesToEntities: ConfigEntryToEntities[] = lockConfigEntries.map(
-      (configEntry) => {
-        return {
-          configEntry,
-          entities: lockEntityEntries.filter(
-            (entity) => configEntry.entry_id === entity.config_entry_id
-          )
-        };
-      }
-    );
-
-    return {
-      title: 'Lock Code Manager',
-      views: configEntriesToEntities.map((configEntryToEntities) => {
-        return {
-          strategy: {
-            type: 'lock-code-manager',
-            ...configEntryToEntities
-          }
-        };
-      })
-    };
-  }
-}
-
-class LockCodeManagerViewStrategy extends ReactiveElement {
+// @customElement('ll-strategy-view-lock-code-manager')
+export class LockCodeManagerViewStrategy extends ReactiveElement {
   static async generate(config: LockCodeManagerViewStrategyConfig, hass: HomeAssistant) {
     const { configEntry, entities } = config.strategy;
 
@@ -130,10 +84,6 @@ class LockCodeManagerViewStrategy extends ReactiveElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    'll-strategy-dashboard-lock-code-manager': LockCodeManagerDashboard;
     'll-strategy-view-lock-code-manager': LockCodeManagerViewStrategy;
   }
 }
-
-customElements.define('ll-strategy-dashboard-lock-code-manager', LockCodeManagerDashboard);
-customElements.define('ll-strategy-view-lock-code-manager', LockCodeManagerViewStrategy);
