+++ rollup.config.js
@@ -6,7 +6,7 @@ import typescript from "@rollup/plugin-typescript";
 const dev = process.env.ROLLUP_WATCH;
 
 export default {
-  input: "custom_components/lock_code_manager/strategy/src/lock-code-manager-strategy.ts",
+  input: "custom_components/lock_code_manager/strategy/src/main.ts",
   output: {
     file: "custom_components/lock_code_manager/strategy/lock-code-manager-strategy.js",
     format: "es",
